clean up `DataArray file`
