+++ src/array/ops/apply.rs
@@ -1,7 +1,7 @@
 use arrow2::array::PrimitiveArray;
 
 use crate::{
-    array::data_array::{BaseArray, DataArray},
+    array::{BaseArray, DataArray},
     datatypes::DaftNumericType,
 };
